ai_summary_service/requirements.txt
fastapi==0.119.0
uvicorn[standard]==0.37.0
httpx==0.27.2
pydantic==2.12.1
pydantic-settings==2.0.3
python-dotenv==1.1.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
redis==5.2.0
google-generativeai==0.3.2
prometheus-client==0.20.0
